introduce SemiTree
rationale:
originally I just wanted to use ThinTree for most stuff,
but self-references make it much less usable, so I
instead resort to deciding for every file, that
it gets "thinned" if it does not contain any references
to currently generated outputs, otherwise it stays
inline. This is useful because we can't share exact
copies of files anyways if they contain self-references.